move multi-boot variables to corresponding header
